requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
gofile-client==1.0.1
idna==3.2
requests==2.26.0
soupsieve==2.2.1
urllib3==1.26.7
